+++ api/src/main/java/io/kafbat/ui/serdes/builtin/mm2/MirrorMakerSerde.java
@@ -1,46 +1,19 @@
 package io.kafbat.ui.serdes.builtin.mm2;
 
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.databind.module.SimpleModule;
 import io.kafbat.ui.serde.api.DeserializeResult;
 import io.kafbat.ui.serde.api.SchemaDescription;
 import io.kafbat.ui.serde.api.Serde;
 import io.kafbat.ui.serdes.BuiltInSerde;
-import java.io.IOException;
+import io.kafbat.ui.serdes.builtin.StructSerde;
 import java.nio.ByteBuffer;
 import java.util.Map;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
-import org.apache.kafka.common.protocol.types.BoundField;
 import org.apache.kafka.common.protocol.types.Schema;
 import org.apache.kafka.common.protocol.types.Struct;
 
 @RequiredArgsConstructor
-abstract class MirrorMakerSerde implements BuiltInSerde {
-
-  protected static final JsonMapper JSON_MAPPER = createMapper();
-
-  protected static JsonMapper createMapper() {
-    var module = new SimpleModule();
-    module.addSerializer(Struct.class, new JsonSerializer<>() {
-      @Override
-      public void serialize(Struct value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
-        gen.writeStartObject();
-        for (BoundField field : value.schema().fields()) {
-          var fieldVal = value.get(field);
-          gen.writeObjectField(field.def.name, fieldVal);
-        }
-        gen.writeEndObject();
-      }
-    });
-    var mapper = new JsonMapper();
-    mapper.registerModule(module);
-    return mapper;
-  }
+abstract class MirrorMakerSerde extends StructSerde implements BuiltInSerde {
 
   protected final boolean versioned;
 
@@ -67,11 +40,6 @@ abstract class MirrorMakerSerde implements BuiltInSerde {
   @Override
   public Serde.Serializer serializer(String topic, Serde.Target type) {
     throw new UnsupportedOperationException();
-  }
-
-  @SneakyThrows
-  protected static String toJson(Struct s) {
-    return JSON_MAPPER.writeValueAsString(s);
   }
 
   @Override
